erson v. 3

+++ LNE_Security/LNE_Security/Person/Address.cs
@@ -7,7 +7,7 @@ namespace LNE_Security
 {
     public class Address
     {
-        public UInt16 ID { get; set; } //TODO: Er denne nødvendig?
+        public int ID { get; set; } //TODO: Er denne nødvendig?
         public string StreetName { get; set; }
         public string HouseNumber { get; set; }
         public UInt16 ZipCode { get; set; }
